Server side authorization (#14)
* Add not found page

* Only show courses you are taking or instructing.

* Only show events for your own courses

* Announcements security

* Add module security

* Fix module authorization and disable anon access.

* Add content security

* Only show courses and modules based on start/end dates.

* Disable the ability to change your own user role

* Prevent downgrading users to "user" level if they are a course instructor

* Fix announcements sort

+++ LearnHub.Server/Controllers/UserController.cs
@@ -11,37 +11,36 @@ namespace LearnHub.Server.Controllers
 	public class UserController : ControllerBase
 	{
 		private readonly IUserRepository _userRepository;
+		private readonly IAuthenticatedUserHelper _userHelper;
 
 
 		[HttpGet("all")]
-		[Authorize(AuthPolicies.AdminPolicy)]
+		[Authorize(AuthPolicies.InstructorPolicy)]
 		[ProducesResponseType(StatusCodes.Status200OK)]
 		public async Task<IActionResult> GetUsers()
 		{
 			return Ok(await _userRepository.GetAllAsync());
 		}
 
-
-		//[HttpPut]
-		//[ProducesResponseType(StatusCodes.Status200OK)]
-		//[ProducesResponseType(StatusCodes.Status404NotFound)]
-		//public async Task<IActionResult> UpdateUserPassword([FromBody] string password)
-		//{
-
-		//	if (await _userRepository.UpdatePasswordAsync(id, password))
-		//	{
-		//		return Ok();
-		//	}
-
-		//	return NotFound();
-		//}
-
 		[HttpPut()]
+		[Authorize(AuthPolicies.AdminPolicy)]
 		[ProducesResponseType(StatusCodes.Status200OK)]
 		[ProducesResponseType(StatusCodes.Status400BadRequest)]
 		[ProducesResponseType(StatusCodes.Status404NotFound)]
 		public async Task<IActionResult> UpdateUserRole([FromBody] UserInfoDto user)
 		{
+			int userId = _userHelper.GetUserId(User);
+
+			// User may not modify themselves
+			if (userId == user.Id)
+				return BadRequest();
+
+			// Do not allow downgrade to user if they are currently a course instructor
+			if (user.Role == RoleDto.User && await _userRepository.IsCourseInstructorAsync(user.Id))
+			{
+				return BadRequest();
+			}
+
 			if (await _userRepository.UpdateRoleAsync(user.Id, user.Role))
 			{
 				return Ok();
@@ -56,6 +55,12 @@ namespace LearnHub.Server.Controllers
 		[ProducesResponseType(StatusCodes.Status404NotFound)]
 		public async Task<IActionResult> DeleteUser(int id)
 		{
+			int userId = _userHelper.GetUserId(User);
+
+			// User may not delete themselves
+			if (userId == id)
+				return BadRequest();
+
 			if (await _userRepository.DeleteAsync(id))
 			{
 				return Ok();
@@ -65,7 +70,7 @@ namespace LearnHub.Server.Controllers
 		}
 
 		[HttpGet("role/{role}")]
-		[Authorize(AuthPolicies.AdminPolicy)]
+		[Authorize(AuthPolicies.InstructorPolicy)]
 		[ProducesResponseType(StatusCodes.Status200OK)]
 		public async Task<IActionResult> GetUsersByRole(RoleDto role)
 		{
@@ -73,9 +78,10 @@ namespace LearnHub.Server.Controllers
 		}
 
 
-		public UserController(IUserRepository userRepository)
+		public UserController(IUserRepository userRepository, IAuthenticatedUserHelper userHelper)
 		{
 			_userRepository = userRepository;
+			_userHelper = userHelper;
 		}
 	}
 }
